Removed bad nebula files, replaced one bad one with original, added another.
The replaced nebula and new nebula both come from the FreeOrion
project, which it seems many of the backgrounds originally came from.
Not all of the bad backgrounds were to be found there, but one was,
and the version they have is better than the version we had. Also
found another nebula background from FreeOrion that should slide
neatly into Naikari.