Edited laser and laser2 to better fit how they're supposed to look.
When laser.webp was exported, it became way more desaturated than
the original on the color part, and it became something more like
an orange color. I eyeballed it, and the result is still a bit less
saturated than it's really supposed to be, but it's much closer.
Also replaced laser2.webp with a similarly edited copy of laser.webp
(since originally, they were the same but just with different
coloration, and laser2 just was never re-exported for whatever
reason).