Removed several low-quality, unused portraits, demoted middleaged to generic.
These portraits were mostly used in the past for missions that have
since been removed from Naikari, or were never in Naikari. Most of
them were crap, but we had recently redone the "Middle-aged Man"
portrait, so decided to repurpose that as a replacement for male5
(one of the particularly bad generic portraits) rather than getting
rid of it.